Add logic to put seedlings back once a seedling item is deleted or order is cancelled

+++ src/controllers/seedlingsOrders.ts
@@ -14,11 +14,15 @@ const addOrderItems: RequestHandler = async (req, res) => {
   const orderItems = req.body as AddSeedlingOrderItem[]
 
   try {
+    /**
+     * A new seedling order item was added
+     * decrement the bench value from seedlingBench
+     */
     const transactionResults = await prisma.$transaction([
-      ...orderItems.map(oI =>
+      ...orderItems.map(orderItem =>
         prisma.seedlingBench.update({
-          where: { id: oI.seedlingBenchId },
-          data: { quantity: { decrement: oI.quantity } }
+          where: { id: orderItem.seedlingBenchId },
+          data: { quantity: { decrement: orderItem.quantity } }
         })
       ),
       prisma.order.update({
@@ -43,13 +47,24 @@ const deleteOrderItems: RequestHandler = async (req, res) => {
   }
   const orderItemId = req.params.orderItemId as unknown as number
   try {
-    const orderItem = await prisma.seedlingBenchOrderItem.delete({
-      where: { id: orderItemId }
-    })
+    const orderItem = await prisma.seedlingBenchOrderItem.findUnique({ where: { id: orderItemId } })
+    if (!orderItem) return new Error('no_order_item')
 
-    //TODO add logic to put benches values back
+    /**
+     * A new seedling order item was deleted
+     * increment the bench value from seedlingBench
+     */
+    const transactionResults = await prisma.$transaction([
+      prisma.seedlingBench.update({
+        where: { id: orderItem.seedlingBenchId },
+        data: { quantity: { increment: orderItem.quantity } }
+      }),
+      prisma.seedlingBenchOrderItem.delete({
+        where: { id: orderItemId }
+      })
+    ])
 
-    res.status(200).json(orderItem)
+    res.status(200).json(transactionResults.pop())
   } catch (e) {
     console.log(e)
     res.status(e.statusCode || 500).json(e)
